LayoutTests/imported/w3c/web-platform-tests/service-workers/cache-storage/window/cache-delete.https-expected.txt


PASS Cache.delete with no arguments 
PASS Cache.delete called with a string URL 
PASS Cache.delete called with a Request object 
FAIL Cache.delete called with a HEAD request assert_equals: Cache.delete should leave non-matching response in the cache. Attributes differ: status. expected 200 but got 0
FAIL Cache.delete supports ignoreVary assert_false: Cache.delete should not delete if vary does not match unless ignoreVary is true expected false got true
PASS Cache.delete with a non-existent entry 
FAIL Cache.delete with ignoreSearch option (request with search parameters) assert_equals: undefined : object[0] Attributes differ: status. expected 200 but got 0
FAIL Cache.delete with ignoreSearch option (when it is specified as false) assert_equals: undefined : object[0] Attributes differ: status. expected 200 but got 0
